fix(banco-central): add fallback to get bank list
